+++ lib/BIGSdb/SubmitPage.pm
@@ -804,7 +804,7 @@ sub _get_isolate_submissions_for_curation {
 			qq(<tr class="td$td"><td><a href="$self->{'system'}->{'script_name'}?db=$self->{'instance'}&amp;)
 		  . qq(page=submit&amp;submission_id=$submission->{'id'}&amp;curate=1">$submission->{'id'}</a></td>)
 		  . qq(<td>$submission->{'date_submitted'}</td><td>$submission->{'datestamp'}</td><td>$submitter_string</td>)
-		  . qq(<td>$isolate_count</td><td>$submission->{'quality'}</td>);
+		  . qq(<td>$isolate_count</td><td>$submission->{'quality'}</td><td>$submission->{'warning_reasons'}</td>);
 		if ( $self->{'system'}->{'dbtype'} eq 'isolates' && $embargo->{'embargo_enabled'} ) {
 			my $embargo_months = $submission->{'embargo'} // '-';
 			$buffer .= qq(<td>$embargo_months</td>);
@@ -828,7 +828,8 @@ sub _get_isolate_submissions_for_curation {
 			  if -e $isolate_curate_message;
 		}
 		$return_buffer .= q(<div class="scrollable"><table class="resultstable"><tr><th>Submission id</th>)
-		  . q(<th>Submitted</th><th>Updated</th><th>Submitter</th><th>Isolates</th><th>Quality</th>);
+		  . q(<th>Submitted</th><th>Updated</th><th>Submitter</th><th>Isolates</th><th>Quality</th>)
+		  . q(<th>Warning reasons</th>);
 		$return_buffer .= q(<th>Embargo requested (months)</th>)
 		  if $self->{'system'}->{'dbtype'} eq 'isolates' && $embargo->{'embargo_enabled'};
 		$return_buffer .= q(<th>Outcome</th>) if $status eq 'closed';